Split the javascript into blacklight and rialto components
This makes it easer to extract blacklight to a standalone library
eventually.
