+++ app/javascript/rialto/layouts/collaborationReport.vue
@@ -1,6 +1,6 @@
 <template>
   <section class="container">
-    <h1>Research Trends Report</h1>
+    <h1>Collaboration Report</h1>
     <label for="school">School: </label>
     <select name="school" v-model="selectedSchool" @change="selectedDepartment=''">
       <option v-for="school in schools" :value="school">{{ school.label }}</option>
@@ -9,24 +9,35 @@
     <select name="department" v-model="selectedDepartment" @change="selectedSchool=''">
       <option v-for="department in departments" :value="department">{{ department.label }}</option>
     </select><br />
-    <label for="yearsRange">Years: </label>
+    <label for="reportType">Report type: </label>
+    <select name="reportType" v-model="selectedReportType">
+        <option value="coauthors">Co-authors</option>
+        <option value="coauthor-institutions">Co-author institutions</option>
+        <option value="coauthor-countries">Co-author countries</option>
+    </select><br />
+
     <YearSlider v-model="selectedYearsRange"></YearSlider>
     <ul v-if="reportURL">
       <li><a href="#" v-on:click="download">Download</a></li>
     </ul>
+    <div v-show="selectedReportType === 'coauthor-countries'">
+      <Choropleth v-bind:reportURL="reportURL"/><br />
+    </div>
     <ReportTable v-bind:data-source="reportURL"></ReportTable>
   </section>
 </template>
 
 
 <script>
-import ReportTable from 'blacklight/reports/table.vue'
-import YearSlider from '../reports/yearSlider'
+import ReportTable from 'rialto/reports/table'
+import Choropleth from 'rialto/choropleth'
+import YearSlider from 'rialto/reports/yearSlider'
 
 export default {
   components: {
+    YearSlider,
     ReportTable,
-    YearSlider
+    Choropleth
   },
   data: function () {
     return {
@@ -34,6 +45,7 @@ export default {
       schools: [],
       selectedDepartment: '',
       departments: [],
+      selectedReportType: 'coauthors',
       selectedYearsRange: [2000, (new Date()).getFullYear()]
     }
   },
@@ -52,10 +64,10 @@ export default {
   computed: {
     reportURL: function(){
       // Must have a report type
-      if (!(this.selectedSchool || this.selectedDepartment)) {
+      if (!this.selectedReportType || !(this.selectedSchool || this.selectedDepartment)) {
           return null
       }
-      return `/reports/research-trends.csv?org_uri=${this.selectedDepartment.uri || this.selectedSchool.uri}&start_year=${this.selectedYearsRange[0]}&end_year=${this.selectedYearsRange[1]}`
+      return `/reports/${this.selectedReportType}.csv?org_uri=${this.selectedDepartment.uri || this.selectedSchool.uri}&start_year=${this.selectedYearsRange[0]}&end_year=${this.selectedYearsRange[1]}`
     }
   },
   methods: {
