+++ app/javascript/rialto/reports/table.vue
