+++ app/javascript/rialto/reports/yearSlider.vue
