+++ app/javascript/test/blacklight/pagination.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 import { shallowMount } from '@vue/test-utils'
-import Pagination from '../blacklight/search/pagination.vue'
-import Result from '../blacklight/search/result.js'
+import Pagination from 'blacklight/search/pagination'
+import Result from 'blacklight/search/result'
 
 describe('result', () => {
   it("shows nothing while loading", () => {
