+++ app/views/layouts/blacklight/base.html.erb
@@ -10,8 +10,8 @@
 
     <title>Rialto</title>
     <%= favicon_link_tag %>
-    <%= stylesheet_pack_tag 'blacklight_vue' %>
-    <%= javascript_pack_tag 'blacklight_vue' %>
+    <%= stylesheet_pack_tag 'rialto' %>
+    <%= javascript_pack_tag 'rialto' %>
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/solid.css" integrity="sha384-VGP9aw4WtGH/uPAOseYxZ+Vz/vaTb1ehm1bwx92Fm8dTrE+3boLfF1SpAtB1z7HW" crossorigin="anonymous">
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/fontawesome.css" integrity="sha384-1rquJLNOM3ijoueaaeS5m+McXPJCGdr5HcA03/VHXxcp2kX2sUrQDmFc3jR5i/C7" crossorigin="anonymous">
     <link href="https://fonts.googleapis.com/css?family=Source+Serif+Pro" rel="stylesheet">
